+++ src/pages/index.ts
@@ -4,4 +4,4 @@ export { default as CoinDrawPage } from './CoinDrawPage'
 export { default as OauthPage } from './OauthPage'
 export { default as MyPage } from './MyPage'
 export { default as StorePage } from './StorePage'
-export { default as PurchasePage } from './PurchasePage'
+export { default as PurchasePage } from './StorePage/components/PurchasePage'
